pyproject.toml
[tool.poetry]
name = "isamples_release_tagger"
version = "0.1"
description = "Python implementation of iSamples release process"
authors = ["Danny Mandel <danny.mandel@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.9"
click = "^7.1.2"
ansible = "^5"
GitPython = "^3.1"